Data loading and setup (#79)
* FEAT: add CSV generation pipeline for auth restaurant booking

* CHORE: add generated processed CSV seed outputs

* DOCS: add setup data generation and loading commands

* CHORE: add AUTH_POSTGRES_DB to env template

* FEAT: update booking and restaurant schema constraints

* CHORE: update ignore rules for data and cache

* REFACTOR: split service-specific CSV logic and improve streaming paths

* FIX: resolve generator mismatch in CSV pipeline

* PERF: speed up reviews loading and category indexing

* PERF: stream restaurant generation directly from CSV files without DataFrames

* DEL: processed data

* CHORE: renamed the files to bug fix

* DATA: Use source gPlus IDs for users and restaurants

* UPD: Ignore zip artifacts

* CHORE: add create_env workflow and simplify env template

* DATA: Replace booking CSV generation with review pipeline and align seed auth password hash

* DATA: Add load-reviews target and switch seed IDs to DB-safe integers

* CHORE: update the makefile to load data into container

* DATA: Expand test coverage to all microservices and verify system validation

* DATA: Align CSV IDs with DB schema across services

* CHORE: Simplify README setup instructions

* FIX: README.md

* FIX: Generate user_id on auth register insert

* FIX: Generate restaurant_id on restaurant register insert

* DATA: Hash gPlus place and user IDs to BIGINT in scripts

* DATA: Migrate service schemas from NUMERIC IDs to BIGINT

* DATA: Document hash-based ID strategy

* DOCS: Document scripts process structure

* DOCS: Add ID strategy heading

* FIX: Migrate user and restaurant IDs to int64 across services

* fix: updated owner id to bigInt

* chore: added '-v' tag to down command in Makefile

* test: added and updated collections for restarant service

* Fix: Change review_id from int32 to int64

* fix: restaurant details connection to reviews

* chore: added more make commands

* chore: removed exposes in docker compose files of all services

---------

+++ services/sponsor/internal/service/service.go
@@ -19,12 +19,12 @@ func New(repo *repository.Repository, client pb.RestaurantToSponsorServiceClient
 }
 
 type sponsorRepository interface {
-	GetRestaurantSponsorship(ctx context.Context, id int32) (*models.SponsorshipResponse, error)
+	GetRestaurantSponsorship(ctx context.Context, id int64) (*models.SponsorshipResponse, error)
 	Sponsor(ctx context.Context, request *models.Sponsorship) (*models.SponsorshipResponse, error)
 }
 
-func (s *Service) Sponsor(ctx context.Context, request *models.Sponsorship, owner int) (*models.SponsorshipResponse, error) {
-	restaurant, err := s.getRestaurantDetails(ctx, int32(request.ID))
+func (s *Service) Sponsor(ctx context.Context, request *models.Sponsorship, owner int64) (*models.SponsorshipResponse, error) {
+	restaurant, err := s.getRestaurantDetails(ctx, request.ID)
 	if err != nil {
 		return nil, err
 	}
@@ -40,11 +40,11 @@ func (s *Service) Sponsor(ctx context.Context, request *models.Sponsorship, owne
 	return s.repo.Sponsor(ctx, request)
 }
 
-func (s *Service) GetRestaurantSponsorship(ctx context.Context, id int32) (*models.SponsorshipResponse, error) {
+func (s *Service) GetRestaurantSponsorship(ctx context.Context, id int64) (*models.SponsorshipResponse, error) {
 	return s.repo.GetRestaurantSponsorship(ctx, id)
 }
 
-func (s *Service) getRestaurantDetails(ctx context.Context, restaurantID int32) (*models.RestaurantDetails, error) {
+func (s *Service) getRestaurantDetails(ctx context.Context, restaurantID int64) (*models.RestaurantDetails, error) {
 	res, err := s.client.GetRestaurantSponsorship(ctx, &pb.GetRestaurantSponsorshipRequest{
 		Id: restaurantID,
 	})
@@ -54,8 +54,8 @@ func (s *Service) getRestaurantDetails(ctx context.Context, restaurantID int32) 
 	}
 
 	return &models.RestaurantDetails{
-		ID:         int(res.Id),
+		ID:         res.Id,
 		Categories: res.Categories,
-		OwnerID:    int(res.OwnerId),
+		OwnerID:    res.OwnerId,
 	}, nil
 }
